deps: bumped all to latest
- fix: in_re_list was reading .source off the filename string instead of the compiled regex
- fix: removed the dead if (!this.list) blocks
- fix: removed the dead data_any guard checking, the hook is only registered when cfg.check.any is true.
