+++ .github/workflows/ci.yml
@@ -3,6 +3,8 @@ name: CI
 on:
   pull_request:
     types: [opened, synchronize, reopened]
+  push:
+    branches: [master]
 
 env:
   CI: true
@@ -10,11 +12,34 @@ env:
 jobs:
   lint:
     uses: haraka/.github/.github/workflows/lint.yml@master
+    permissions:
+      contents: read
+
+  prettier:
+    if: ${{ github.event_name == 'push' || github.event.pull_request.head.repo.full_name == github.repository }}
+    uses: haraka/.github/.github/workflows/prettier.yml@master
+    permissions:
+      contents: write
+    with:
+      branch: ${{ github.event_name == 'pull_request' && github.head_ref || github.ref_name }}
+
+  coverage:
+    uses: haraka/.github/.github/workflows/coverage.yml@master
+    secrets: inherit
+    permissions:
+      contents: read
 
   ubuntu:
-    needs: [lint]
     uses: haraka/.github/.github/workflows/ubuntu.yml@master
+    permissions:
+      contents: read
 
   windows:
-    needs: [lint]
     uses: haraka/.github/.github/workflows/windows.yml@master
+    permissions:
+      contents: read
+
+  macos:
+    uses: haraka/.github/.github/workflows/macos.yml@master
+    permissions:
+      contents: read
