+++ .github/workflows/release.yml
@@ -1,0 +1,27 @@
+name: release
+
+on:
+  push:
+    branches: [master, main]
+    paths:
+      - package.json
+
+env:
+  CI: true
+
+jobs:
+  release:
+    uses: haraka/.github/.github/workflows/release.yml@master
+    secrets: inherit
+    permissions:
+      contents: write
+
+  publish:
+    needs: release
+    if: needs.release.outputs.changed == 'true'
+    uses: haraka/.github/.github/workflows/publish.yml@master
+    secrets: inherit
+    permissions:
+      contents: read
+      packages: write
+      id-token: write
