+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "haraka-plugin-access",
-  "version": "1.1.10",
+  "version": "1.2.0",
   "description": "Haraka plugin for ACLs for email connections",
   "main": "index.js",
   "files": [
@@ -13,9 +13,11 @@
     "lint:fix": "npx eslint *.js test --fix",
     "prettier": "npx prettier . --check",
     "prettier:fix": "npx prettier . --write --log-level=warn",
-    "test": "npx mocha",
-    "versions": "npx dependency-version-checker check",
-    "versions:fix": "npx dependency-version-checker update && npx prettier package.json --write --log-level=warn"
+    "test": "node --test",
+    "test:coverage": "node --test --experimental-test-coverage",
+    "test:coverage:lcov": "mkdir -p coverage && node --test --experimental-test-coverage --test-reporter=lcov --test-reporter-destination=coverage/lcov.info",
+    "versions": "npx npm-dep-mgr check",
+    "versions:fix": "npx npm-dep-mgr update"
   },
   "repository": {
     "type": "git",
@@ -33,14 +35,14 @@
   },
   "homepage": "https://github.com/haraka/haraka-plugin-access#readme",
   "devDependencies": {
-    "@haraka/eslint-config": "^1.1.5",
-    "haraka-test-fixtures": "^1.3.8"
+    "@haraka/eslint-config": "^2.0.4",
+    "haraka-test-fixtures": "^1.4.3"
   },
   "dependencies": {
-    "address-rfc2822": "^2.2.2",
-    "haraka-net-utils": "^1.7.0",
-    "haraka-tld": "^1.2.2",
-    "haraka-utils": "^1.1.3"
+    "address-rfc2822": "^2.2.3",
+    "haraka-net-utils": "^1.8.2",
+    "haraka-tld": "^1.3.3",
+    "haraka-utils": "^1.1.4"
   },
   "prettier": {
     "singleQuote": true,
